UNIONVMS-935 List user defined areas
[I had to change all the SYS areas entities to map to Long ID instead to Int, due to a common API method with User areas - which are already using Long]

git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Spatial/APP@7588 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/spatial/dao/CommonDao.java
@@ -90,9 +90,9 @@ public abstract class CommonDao {
 		return query;
 	}
 
-	protected Query createNamedQueryWithParameterList(String nativeQuery, Map<String, List<Integer>> parameters) {
+	protected Query createNamedQueryWithParameterList(String nativeQuery, Map<String, List<Long>> parameters) {
 		Query query = getSession().getNamedQuery(nativeQuery);
-		for (Map.Entry<String, List<Integer>> entry : parameters.entrySet()) {
+		for (Map.Entry<String, List<Long>> entry : parameters.entrySet()) {
 			query.setParameterList(entry.getKey(), entry.getValue());
 		}
 		return query;
