+++ service/src/main/java/eu/europa/ec/fisheries/uvms/spatial/entity/EezEntity.java
@@ -41,7 +41,7 @@ public class EezEntity implements Serializable {
 	@Column(name = "gid")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@ColumnAliasName(aliasName ="gid")
-	private int gid;
+	private long gid;
 
     @Basic
     @Column(name = "geom")
@@ -124,11 +124,11 @@ public class EezEntity implements Serializable {
     public EezEntity() {
 	}
 
-	public int getGid() {
+	public long getGid() {
 		return this.gid;
 	}
 
-	public void setGid(int gid) {
+	public void setGid(long gid) {
 		this.gid = gid;
 	}
