+++ service/src/main/java/eu/europa/ec/fisheries/uvms/spatial/entity/RfmoEntity.java
@@ -40,7 +40,7 @@ public class RfmoEntity implements Serializable {
 	@Column(name = "gid")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@ColumnAliasName(aliasName="gid")
-	private int gid;
+	private long gid;
 	
     @Basic
     @Column(name = "geom")
@@ -63,11 +63,11 @@ public class RfmoEntity implements Serializable {
 	public RfmoEntity() {
 	}
 
-	public int getGid() {
+	public long getGid() {
 		return this.gid;
 	}
 
-	public void setGid(int gid) {
+	public void setGid(long gid) {
 		this.gid = gid;
 	}
