+++ service/src/main/java/eu/europa/ec/fisheries/uvms/spatial/entity/StatRectEntity.java
@@ -33,16 +33,16 @@ public class StatRectEntity implements Serializable {
 	@Column(name = "gid")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@ColumnAliasName(aliasName="gid")
-	private int gid;
+	private long gid;
 
 	public StatRectEntity() {
 	}
 
-	public int getGid() {
+	public long getGid() {
 		return this.gid;
 	}
 
-	public void setGid(int gid) {
+	public void setGid(long gid) {
 		this.gid = gid;
 	}
